Update README.md
Note this Web app is exclusively for ALX Portfolio Project Defense




Introduction of  Zora Jobs
Zora jobs is a marketplace software tool designed to help recruiters who seek for FrontEnd Developers, to feel up roles in their businesses. 
Its an application that houses just frontend developers, where their github account can be accessible for all, its know as a hiring company. 

Basic Feature of of the App
    The app, is known for displaying a slideshow of job seekers who are developers on its resume portals, for potential jobs.
Advanced Features
It has buildups to monitor site visits, updates of its members, and tracks recruited developers development with is registration numbers, it helps intending employers to engage developers of their choices access to github accounts and as well as connecting  them. 

Benefits of Using This App
 It benefits young developers with or without prior working experience the opportunity of landing their first job roles,
It helps them get the engagement the deserve and help them build a proper resume for themselves.
It's a marketplace for developers to come together and contribute to open source projects.

Github Repository 
The Github repository includes the code,
And documentation of the application.


Webstack - Portfolio Project

Technologies Used
-HTML
-CSS
REACT.JS
JAVASCRIPT



                                    Reports Of Development
Success
In this first phase, Zora Jobs has extensively covered its primary needs, which is where developers see to be a place that helps them get hired, and helps  build up their resumes for jobs.
  Its an app that has successfully been used by developers to get to their peers around the globe be building things together.
  2.   Failures 
This app in this phase, can not implement API searches that will filter developers by locations and cites. 

Future Plans For The App
My plans are to add more functionality, and expand its scope to meet day to day activities of the end users, to help the users with options of recruiting from their tribe, religion and nationality. 
Summary/Conclusion 
 I plan on adding Backend developers the opportunity in our app, and as well as having weekly tasks among developers to bring out the best at all times for jobes.
This will not only benefit the users of the app, but also provide a valuable learning opportunity for developers who are interested in contributing to open source projects. Therefore, I am excited to share this project with the world and look forward to seeing how it can evolve with the help of the open source community.

Here is the Demo of the Project,
https://jayteezany.github.io/Jayteezany-Webstack_Portfolio_Project-/